Correction of minor bugs

+++ internal/pool/page.go
@@ -1,6 +1,6 @@
 package pool
 
-var PageSize uint16 = 4096
+var PageSize uint64 = 4096
 
 // Page is the unit of the Bufferpool
